+++ example/android/app/src/main/kotlin/com/example/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.example
+package asap_ead.plataformaasap.com.br
 
 import io.flutter.embedding.android.FlutterActivity