+++ lib/check_version.dart
@@ -1,3 +1,4 @@
 library check_version;
 
 export 'src/check_version.dart';
+export 'src/model/version_status.dart';